- add version validation and update to 20130329 (which fails that validation unfortunately; upstream mistake)

+++ youtube-dl.spec
@@ -1,11 +1,12 @@
+%define	ver	2013.03.29
 Summary:	Video extraction utility for YouTube
 Summary(pl.UTF-8):	Narzędzie do wydobywania filmów z YouTube
 Name:		youtube-dl
-Version:	20130329
+Version:	%(echo %{ver} | tr -d .)
 Release:	1
 License:	Public Domain
 Group:		Applications/System
-Source0:	https://github.com/rg3/youtube-dl/raw/2013.03.29/youtube-dl
+Source0:	https://github.com/rg3/youtube-dl/raw/%{ver}/youtube-dl?/%{name}-%{version}
 # Source0-md5:	a2a30e4b9ad06538d92968ceb2d58b3f
 URL:		http://rg3.github.com/youtube-dl/
 BuildRequires:	rpm-pythonprov
@@ -23,6 +24,15 @@ YouTube.com.
 youtube-dl jest programem do ściągania plików video z YouTube.com.
 
 %prep
+%setup -c -T
+
+%build
+ver=$(%{__python} %{SOURCE0} --version | tr -d .)
+
+if [ "$ver" != "%{version}" ]; then
+	echo "Source with $ver found while expecting %{version}!" >&2
+	exit 1
+fi
 
 %install
 rm -rf $RPM_BUILD_ROOT
